fix(cli): panic-safe env restore for HOME/STELLA_DATA_DIR tests (#911)
42 tests across 12 files mutated HOME/STELLA_DATA_DIR then restored
them by hand at the end of the function. If an assertion panicked
mid-test, the manual restore never ran, leaving the real value
corrupted for every test that runs afterward in the same binary.

Promotes EnvRestore (already proven in enterprise_telemetry_tests.rs)
into the shared test_env module and migrates every other HOME/
STELLA_DATA_DIR-mutating test onto it: capture-on-entry, Drop-based
restore that runs even on unwind, still serialized behind the
existing test_env::lock() mutex. skill_manager's scratch() helper
never restored HOME at all before this — that gap is now closed via
a ScratchGuard that bundles the lock and the restore with a
deliberate field-drop order.

Full suite verified green at --test-threads=1/2/4/8 and default.
