+++ stella-cli/src/settings/private_state_tests.rs
@@ -8,12 +8,12 @@ fn user_settings_are_private_but_project_settings_modes_are_untouched() {
     use std::os::unix::fs::PermissionsExt;
 
     let _env = crate::test_env::lock();
+    let _restore = crate::test_env::EnvRestore::capture(&["HOME"]);
     let tmp = tempfile::tempdir().unwrap();
     let home = tmp.path().join("home");
     let user_dir = home.join(".stella");
     std::fs::create_dir_all(&user_dir).unwrap();
     std::fs::set_permissions(&user_dir, std::fs::Permissions::from_mode(0o777)).unwrap();
-    let previous_home = std::env::var_os("HOME");
     // SAFETY: serialized behind the binary-wide environment lock.
     unsafe { std::env::set_var("HOME", &home) };
 
@@ -40,11 +40,6 @@ fn user_settings_are_private_but_project_settings_modes_are_untouched() {
     engine.save_to(&project).unwrap();
     assert_eq!(mode(&project_dir), 0o777);
     assert_eq!(mode(&project), 0o664);
-
-    match previous_home {
-        Some(previous) => unsafe { std::env::set_var("HOME", previous) },
-        None => unsafe { std::env::remove_var("HOME") },
-    }
 }
 
 #[cfg(unix)]
@@ -53,11 +48,11 @@ fn user_settings_save_rejects_a_symlink_without_touching_target() {
     use std::os::unix::fs::symlink;
 
     let _env = crate::test_env::lock();
+    let _restore = crate::test_env::EnvRestore::capture(&["HOME"]);
     let tmp = tempfile::tempdir().unwrap();
     let home = tmp.path().join("home");
     let user_dir = home.join(".stella");
     std::fs::create_dir_all(&user_dir).unwrap();
-    let previous_home = std::env::var_os("HOME");
     // SAFETY: serialized behind the binary-wide environment lock.
     unsafe { std::env::set_var("HOME", &home) };
     let user = user_settings_path().unwrap();
@@ -67,9 +62,4 @@ fn user_settings_save_rejects_a_symlink_without_touching_target() {
 
     assert!(AgentEngineConfig::default().save_to(&user).is_err());
     assert_eq!(std::fs::read_to_string(&target).unwrap(), "{}\n");
-
-    match previous_home {
-        Some(previous) => unsafe { std::env::set_var("HOME", previous) },
-        None => unsafe { std::env::remove_var("HOME") },
-    }
 }
